Fixed warning in example.

+++ Example/PullToRefreshDemo/ObjcTableViewController.m
@@ -33,7 +33,7 @@
 - (UITableViewCell *)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath {
 	UITableViewCell *cell = [tableView dequeueReusableCellWithIdentifier:@"DemoCell" forIndexPath:indexPath];
 	
-	cell.textLabel.text = [NSString stringWithFormat:@"Row %d", indexPath.row];
+	cell.textLabel.text = [NSString stringWithFormat:@"Row %ld", (long)indexPath.row];
 	
 	return cell;
 }
